src/requirements.txt
psutil
keyboard
pystray
Pillow
